test: update and move integration tests into src folder
